tpu_requirements.txt
tqdm
ipdb
h5py
pyyaml
moviepy
wandb
av
flax
gcsfs==2022.7.1
tensorflow_cpu
tensorflow_datasets
tensorflow_io
tensorflow_gan
tensorflow_probability
tensorflow_hub
lpips_jax
numpy